r/consensus: fixed updating leader commit index
The leader commit index and committed entries notifications has to be
done under the semaphore. Incorrect implementation lead to the deadlock
between two fibers acquiring the raft operation semaphore. (Replicate
entries STM was executed inside of the semaphore and it wasn't able to
finish processing because semaphore was also acquired in
process_append_replies while triggering commit index update). Fix
include update leader commit index in background. The updates to commit
index can not be concurrent but they are commutable between each other.
Therefore the updates can be done in background under semaphore. This
allows the replicate STM to finish and continue processing.

Signed-off-by: Michal Maslanka <michal@vectorized.io>