model: simplify model::record_batch_reader to use concrete records
Record batch reader forced all record_batch_reader::impl classes
to maintain the buffer for the reader to manipulate.

Furthermore, it force the reader to set *internal* variable called
_end_of_stream=false when we've reached the end *in addition to*
also overriding interface methods.

This change makes the record_batch_reader::impl take a
ss:circular_buffer<model::record_batch> which eliminates all
state tracking since we can pop the front and the back.
It has the _same_ performance characterestics as a vector and the
growth is amortized.

With the recent migration in the raft:: to migrate to record_batch_reader
and due to the fact that record_batch_reader has no synchronous
interface, this patch added serialization helpers that defer
*rpc* specific serializers to a light weight wrapper in
reflection::async_adl<T> introduced in previous patches.