kafka: delegate parsing of kafka to iobuf_parer
Make the kafka_reader.h delegate all the calls to the iobuf from
a member iobuf_parser without changing the interface.

Delete all unused methods in the codebase.

This is an infrastructure patch for stabilizing the new storage
format including null keys and null bytes to leverage the same
parser

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200201064048.1782977-1-alex@vectorized.io>