storage: become compatible with kafka below the CRC
In order to properly compute the CRC and detect corruption, both
for the clients and for us, we must store all the fields that
are taken into account by the CRC.

In addition, we must continue to exclude our fields *not* in the
kafka batch CRC calculation.

This patch introduces full parsing of the kafka batch format
including checksuming during decoding to verify that what we
have received is *not* different from what we read from the wire.

In order to minimize code duplication the `make_random_batch`
util was merged into a v::storage_test_util library so that
multiple layers can benefit from the unification which
includes correct (as per kafka) CRC. Which means we convert
cpu_to_be() to compute the CRC.

The disk and parser were also updated to be able to properly parse
and the extended format from disk.

All other subsystems were updated with the cosmetic additions of
of the missing fields.

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200203033214.116155-1-alex@vectorized.io>