storage: bugfix update _flushed_pos = _pos after truncate()
While truncating a segment to a particular offset, (say middle),
we must readback half-page (or some bye offset in the middle of a page)
to be able to continue DMA writing to the file handle.

In order to do this, we exposed a low level api of the
log_segment_appender::chunk to simply set the position, so we wouldn't
have to double-buffer the half page and use the regular append.

The reason is we already have page-aligned buffers sitting in the appender
that we can simply re-use instead of allocating new memory.
In order to correctly indicate the _follow up_ flush, we must also let
the _flushed_pos know that we are indeed all caught up.