vtools: add simple storage summary tool
this is an evolving tool that I have been using to sanity check data
created through the kafka api. some or all of it will probably be
rewritten in c++ at some point, but python makes it very easy to iterate
on the different ways in which such a tool is useful.

Signed-off-by: Noah Watkins <noah@vectorized.io>