r/stms: fixed stop condition in vote and replicate stm
In case when raft group contains even number of nodes the append entries
or vote stm may never finish. For example if one of the nodes vote for
self and the other as well. Then there will be no majority for both
successful and failed votes. The stop condition in state machines that
relay on majority should also check if all nodes responded to request.

Signed-off-by: Michal Maslanka <michal@vectorized.io>