r/replicate_state_stm: waiting for self append
Replicate STM has to wait for self append to happen before returning
control. Raft semaphore is held for the time data are replicated.
Previously if self append was slower then append on the remotes the
background future for self append was executed outside of the semaphore,
moreover leader could have a log that wasn't the most up to date from
all the nodes. In order to prevent that from happening added promise
allowing replicate STM to wait efficiently for leader to finish append
while allowing all the appends (remote and local one) to be executed in
parallel.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20200206115025.282546-1-michal@vectorized.io>