controller: make controller be a sharded service
although all the meaningful controller state and computation occurs on
core0 this patch changes the controller to be a sharded service. this
change is in support of an initially simple cross-core use case within
the controller which is handling leadership notifications. the service
is a well understood abstraction that we reuse to deal with races on
shutdown.

this patch does nothing other than make the switch to being sharded.

the next patch builds upon the sharded service to handle leadership
notifications across cores.

Signed-off-by: Noah Watkins <noah@vectorized.io>