r/types: introduced type used by leadership notification
The leadership notification should be triggered not only when current
node win the voting round but also when other node become a leader of a
cluster. Introduced type leadership_status to bound the group id with
current group leader. Additionally made the leader_id optional to
reflect the situation when there is no leader in the group.

Signed-off-by: Michal Maslanka <michal@vectorized.io>