kafka: bugfix: account for kafka and vectorized's size missmatch
Vectorized's storage format includes *all* bytes in the header.size_bytes
field. Kafka on the other hand *only* includes everything below the
batch-length as the size-bytes.

Adapt the batch-length to our size-bytes transformation in the kafka adapter
as well as move the known value of the current header format to the
model/record.h so it can be reference in multiple places. Deletes the
storage/constants.h which are really the record header overhead

Tests: produce and reover 1000 batches from the kafka python client