kafka: fix request context linking issue
This is a bit crazy. There is a struct op_context defined in
produce_request.cc. There is also a struct op_context defined in
fetch_request.h that is used when handling fetch request.  Both of these
objects have a member named request.decode, but request is of different
type (fetch_request and produce_request).

This code compiles fine.

But this happens when handling a fetch request. It appears to be linking
against the wrong op_context and calling decode as if a produce request
is being handled.

8  0x000055556112fae2 in kafka::produce_request::decode (this=0x7fffee4f91c0, ctx=...) at ../../../src/v/kafka/requests/produce_request.cc:49
9  0x00005555611525cb in kafka::op_context::op_context (this=0x7fffee4f9110, rctx=..., ssg=...) at ../../../src/v/kafka/requests/produce_request.cc:135
10 0x000055556160996f in kafka::fetch_api::process (rctx=..., ssg=...) at ../../../src/v/kafka/requests/fetch_request.cc:372
11 0x0000555561121517 in kafka::do_process<kafka::fetch_api> (ctx=..., g=...) at ../../../src/v/kafka/requests/requests.cc:39
12 0x000055556111c15f in kafka::process_request (ctx=..., g=...) at ../../../src/v/kafka/requests/requests.cc:62

This patch changes the name of one of the op_context types and the
problem goes away.

Signed-off-by: Noah Watkins <noah@vectorized.io>