storage: return next batch from batch cache
There are tree possible results when reading a range from the batch
cache, and these have implications on how the caller should handle
the next offset when scanning the log.

The cases are:

1. empty (none cached)
2. empty (all skipped)
3. non-empty

When (1) occurs the caller should resume scanning the log at the same
offset the cache was queried at. It is a cache miss.

In (2) the caller should resume based on the offset of the last batch,
even though it is filtered out of the result.

Case (3) is identical to case (2) when the last batch in the result had
been filtered.

All of these cases are handled by seeding the log iteration position
from the result `.next_batch` offset.

Signed-off-by: Noah Watkins <noah@vectorized.io>