storage: fix offset account with batch type filtering
When skipping batches because they do not pass the type filter the "next
offset" still needs to be udpated. This patch performs that
filtering-aware offset update, and moves the hole-checking logic down
into the skipping consumer. Leaving it at higher level is problematic
because a hole vs a skipped batch could not be easily recognized.

Signed-off-by: Noah Watkins <noah@vectorized.io>