cluster: unregister controller callback on shutdown
There is a circular dependency between the controller and the partition
manager/raft. The controller has a direct dependency on the partition
manager and the partition manager has a dynamically defined dependency
via registered callbacks. This results in a race between callbacks and
the controller.

This patch adds interfaces for unregistering these callbacks eliminating
the dependency on shutdown.

Signed-off-by: Noah Watkins <noah@vectorized.io>