vtools/infra: Run terraform init preemptively
The "cluster" Terraform module depends on the "aws" and "random" providers,
which are installed by running `terraform init` in the module directory.
This patch makes the `deploy` and `destroy` subcommands run `terraform init`
automatically every time they're executed. `init` is idempotent, and doesn't
reinstall the providers and plugins if they're already installed.