r/consensus: fixed reading configuration
Last seen configuration offset is initialized with model::offset{}. The
reader can not read batches with negative offsets hence the
configuration was never processed causing failure when joining the
cluster. (Configuration was never processed by raft).
This fix advances the last_seen_config_offset as it is inclusive and it
should be read from the next one.

Signed-off-by: Michal Maslanka <michal@vectorized.io>