bugfix:storage/cache: ensure that cached reads do not regress offset
log_reader.cc: blindly assigns the bottom offset of the cache to the
starting offset, as intended by the api via:

```
// handles cases where the type filter skipped batches. see
// batch_cache_index::read for more details.
_config.start_offset = cache_read.next_batch;
```

However, if the cache _must_ always return at least the input offset
of the query for that line to work.

This fixes a bug where consensus was bottlenecking around the
_reads_ for a _write only workload_ pegging around 3MB/s where
the IO for writes should be 0.

Tested by writing 100K with the python client API