vtools/deploy: Make 'destroy' a flag for the cluster cmd
Since we're moving away from Terraform-only deployments, into a more "hybrid"
approach, it makes sense for each deploy subcommand to define how the resources
it deploys should be destroyed. Therefore, its easier and less verbose to model
that as a flag in each subcommand than as a general command, with a subcommand
for each component (like `vtools destroy <component>`).