kafka: check for join group errors before dispatching
This patch adds a static membership (unsupported feature) check before
incurring the cost of dispatching the request to the target core.

Since the last time group membership has been worked on there has been
development within kafka in other areas, and this patch also refactors a
bit to align with patterns for handling requests that emerged with
consume/produce.

Signed-off-by: Noah Watkins <noah@vectorized.io>