kafka: do not handle new flexible version protocol
There are two bugs fixed in this patch.

First, when an API versions request arrives if it is out of bounds of
the max supported it should be treated as if it is version 0 (which
returns all information about APIs). Specifically, we cannot decode it
because we don't know what format the request has if it isn't in the
supported version range.

Second, we do not yet support the new flexible version protocol in
kafka, and the supported bounds allows clients to use this protocol
variant.

Signed-off-by: Noah Watkins <noah@vectorized.io>
Message-Id: <20200218234302.28925-2-noah@vectorized.io>