rpc: abstract connection-level protocol management for reuse
The kafka rpc server and the regular rpc server are near identical,
but our regular RPC has been more battle tested than the kafka
layer and has more extensive unit testing.

This is an infrastructure change and should introduce no
end-user-visible behavior change.

By abstracting the tcp connection management, output stream batching,
TLS connection management, socket listening services, etc we allow
for pluggable protocols which as defined in the interface,
it takes a valid tcp connection and returns a future<>. What it
does with the connection is up to the protocol.

This patch decouples the default built-in protocol in the rpc::server
into a rpc::simple_protocol.