kafka: clean-up and clarify use of group mapper
The group shard mapper that was in place mapped groups to shards.
However, this combined two different, separate operations. The first
operation is mapping group ids to ntps. The second opeartion is mapping
the ntp to a particular shard and partition.

The first operation--mapping groups to ntps--can be done with a
hash-based mapping function (current solution), but the abstraction of a
mapper is left in place to handle future policies that may require
managing explicit mappings. The second operation--mapping ntps to
partitions--is done using the existing shard table and partition manager
sub-systems.

This patch combines both operations in the group router.

This patch also renames the group_shard_mapper to
coordinator_ntp_mapper. First, it no longer takes on the role of mapping
anything to shards. And second the name is changed to coordinator to
reflect the generality of this mechanism w.r.t to the kafka protocol.
The kafka protocol allows more than one type of coordinator resource
(currently groups and transactions). This renaming is simply a step
towards being more unified with the kafka protocol which will reduce
confusing in the future.

Signed-off-by: Noah Watkins <noah@vectorized.io>