cluster: register for partition management callbacks
this patch adds an interface to the partition manager that can be used
to request callbacks when partitions come under management. it uses an
index over the components of an ntp allowing to register for things like
all partitions under a particular topic.

this is initially useful for consumer groups which use a special topic
in which all contained partitions contain group state.

Signed-off-by: Noah Watkins <noah@vectorized.io>