storage: add time-based indexing
These are infrastructure patches needed for time based indexing
to be fronted by the kafka API. Future patches will integrate
with the Kafka API.

We leverage the reflection API to serialize and deserialize the
index to-from disk. We keep the same strategy for offset indexing
which by default stores 'checkpoints' every 32KB of data.

This patch piggy backs to the existing infrastructure for time-based
indexing.

testing: clang,unit