vlog.h: add utility for line number context since c++20 is approved
c++20 adds std::source_location which we will use to add context
to our logging functionality. In the meantime, add a static
constexpr classes with memoization for printing the filename
and line number until we move to c++20.

Once we compile with c++20 vlog will become a template function
and no functionality will be lost.