iobuf: do not limit the size of pre-allocated buffers.
Because we are using seastar input stream which transitively
simply reads what the kernel makes available to the application,
it makes no sense to limit the buffer size to 128KB max, which
actually does _not_ limit the size, but rather simply does a
temporary_buffer<char>::share() under the covers.

This patch eliminates the waste of sharing in case buffers are large.

Signed-off-by: Alexander Gallego <alex@vectorized.io>