c/metadata_cache: storing term in partition metadata
In order to prevent stale leadership updates, the metadata_cache has to
use an epoch. The epoch in current implementation would be the raft
group term. The term has to be stored separately for each partition in
metadata cache to filter stall updates. Added term parameter as an
additional parameter stored per ntp in metadata cache.

Signed-off-by: Michal Maslanka <michal@vectorized.io>