cluster: add cache interface to wait on ntp leadership
this adds an interface that allows for waiting (with timeout) and a
particular ntp in the cache to be updated with a known leader. this is
used in the group membership find coordinator call which may perform the
combined action of creating a new topic and then waiting for a leader to
be elected before replying back to the client. since this can occur on
any node in the cluster, we cannot rely on using a reference to an
actual partition object which also has an interface for waiting on
leadership.

Signed-off-by: Noah Watkins <noah@vectorized.io>