build:bugfix: gcc-9 libstdc++ does not mark std::deque as noexcept
Nor marking the std::deque as noexcept causes move-ctor to be disabled
because it doesn't match the expected value by the compiler.

This in turn causes all future<dequeue<T>> to fail because future<>
requires the types to be move constructible and move-assignment.

Last, we cannot store references because that causes the move-assignment
to fail the noexcept which is required by seastar asserts to work
with future<T>