s/log_replayer: fixed recovering of not truncated segments
In order to achieve best performance we use fallocating file appender.
In normal operating condition when appender is closed the underlying
file is truncated to match the size of data. During crash however we are
unable to truncate the file properly. We have to account for this
situation in log_replayer. Fallocated but not yet written file is filled
with zeros therefore batch all batch header fields will have values set
to 0. Introduced check stopping parsing when batch headers bytes size is
equal to 0.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20200311150136.149405-1-michal@vectorized.io>