kafka:protocol: add connection_context for response sequencing
currently, sequencing was global per shard, this change makes it
local per connection.

This fixes a bug with global connection management. As per discussions
with Noah:

```
This patch manages responses on a per-connection basis (as opposed
to per-shard) which fixes a bug in which responses were sent to
the wrong connection.

`kafka::protocol::_responses` holds responses for all connections
managed on a shard. The responses are indexed per globally unique
sequence number assigned to each request. After any request is
processed `process_next_response` is called to flush any ready
responses in the order that they were created. That is generally
correct behavior.

However, handling the responses didn’t take into account the
correlation between a connection and _its_ responses. Here’s the
pseudo-code for handling responses:

   process_next_response(rpc::server::resources rs)
     while true
        resp = _responses[next]
        next++
        rs.conn->write(resp)

Notice that the next response is taken from the global pool and
sent out over the connection without considering if the response
is for a request that originated on that connection.

This issue is fixed in this patch by managing responses on a per
connection basis making this correlation implicit by structure.

```