c/controller: fixed state recovery when controller become a leader
When controller becomes the leader its state has to be up to date with
the one available in raft-0. In order to make sure that the state is up
to date, after leadership notification is triggered, we
perform write of checkpoint record batch.

When the replication is successful controller has all information
necessary to serve user requests. (i.e topics metadata, partition allocator).

Signed-off-by: Michal Maslanka <michal@vectorized.io>