vtools/pz: adds --rp-extra-conf flag
Adds a flag that allows to specify extra configuration for RP that is
applied by the ducktape service. This configuration extends or overrides
the basic (minimal) configuration implemented by the redpanda ducktape
wrapper.

The flag points to a config file that, when given, it's mounted inside
the ducktape container so that ducktape has it available to it. This
configuration is then applied to all the nodes in the test before
redpanda is started.

In addition, this command now generates the ducktape configuration for
the terraform-deployed cluster, instead of having terraform output
generate it. The SSH key used to connect to these AWS nodes is now the
known test key, since this needs to live inside the container, otherwise
distinct permission (if we bind-mount it from the host) cause the SSH
connection to fail