raft: reduce recovery smt to 32KB chunks.
Currently we try to recover at 1MB for _every_ raft group.
We have no way to control this concurrency and apply backpressure.
Reduce size to 32KB which is a good size batch to get some level
of throughput at every layer. If this setting proves difficult
we can introduce a core-local param (from the log_manager?)
that controls how fast/slow we should be consuming from the log

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200319164304.10098-1-alex@vectorized.io>