rpc:server: always handle exceptional futures in proto apply
When applying a server protocol (user level app), we dispatch
the application on a background future guarded via a gate to wait
on the processing during shutdown. However, shutting down the connection
is not exception safe. Make the protocol application exception safe
to ensure we wait on all background futures, including exceptional
futures. This bug was revealed while testing kafka with 4 parallel
producers and shutting them down at random.