controller: prevent recovery across missmatching core count
If we create topic assignments but restart the service with different
number of cores between each run, we can no longer recover the original
shard assignment. For the purposes of the partition-allocator, this
node looks like a new node. Users can always increase core count,
but they cannot reduce core count.

In a future version we should provide local-reshuffling of groups
and gracefully recover.