iobuf: fix small buffer appends.
This fixes a bug seen in the wild by the kverify tool in our Go
toolbox which sends data of exacltly 1 byte key & value. If each
request was 50K of these 1-byte buffers, we'd effectively
overrun the scattered_message<char> limit of int16_t. Instead,
we now set up guards in iobuf to copy the buffers when they are
below the minimum fragment size.

In addition, this change removes the stateful allocator for a
simple lookup table.