kafka: unify in-memory and on-disk member state
this patch refactors member state so that the on-disk and in-memory
state stores are the same, which simplifies quite a bit of code.

it uses native std::chrono::duration types in serialization rather than
converting to/from integer by hand.

finally it starts to remove the bytes type from the core types used in
group membership but bytes is not yet completely removed. this
introduces some extra copies, but member state should be quite small.

Signed-off-by: Noah Watkins <noah@vectorized.io>