cmake: adds 'Dev' build-type
we need another build-type called 'Dev' that basically is -O1 -g and no
sanitizers. We need this in order to track a memory leak that with
release is not visible due to not having debug symbols, and with `debug`
builds it halts because of address sanitizer's slowness. Thus we need an
"in-the-middle" release to debug some memory leaks.