storage:segment_appender: correctly reset the last half-written-page
We cannot assume that the back of the _free_chunks will always be
the head of the list. We don't know how long IO's will take and
the order the reactor will return the io completions to us.

We must always _find_ the half page, which is characterized by
being the only page that is non-empty since all the otherones would
have been a full reset.