iobuf: delete move ctor/assignment from io_fragment
there are currently no users of these interfaces, but more importantly,
the default implementations are malformed for intrusive lists.

the io_fragment contains a list hook which defines a no-op copy
constructor. thus, when moving from io_fragment the moved from fragment
is unmodified, and the moved to fragment node is initialized as if it is
not on a list.

to safely enable these use the swap_nodes interface.

Signed-off-by: Noah Watkins <noah@vectorized.io>