bytes/iobuf: fixed appending existing iobuf
When iobuf or seastar::temporary_buffer was appended to iobuf its
allocation strategy was bypassed. This caused the iobuf fragment list to
grow beyond the ss::net::packet fragments limit and triggered assertion.
Fixed an error by deciding whether the appended buffer is able to fit in
the next allocation unit, if this is the case the new fragment is
allocated and buffer is copied into it.

Signed-off-by: Michal Maslanka <michal@vectorized.io>