iobuf: expose raw fragment append for io_foreign_share+friends
The new scheme copies the contents of the temporary buffers
which is a fix for small payloads that will overwhelm a
scattered_message<char> maximum size (int16_t). The fix for that
real-world use case caused an unforseen side effect which
is we now 'copy' most of the share().

This patch exposes ownership semantics for iobufs that
do not copy and simply take ownership of fragments.

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200327191036.828027-1-alex@vectorized.io>