bugfix:storage:cache: Always copy memory into the batch cache
Temporary buffers, may, or may not refer to the memory they claim
to be refering. That is, they may at any point hold a deleter
to a *much* larger memory pool than they think they own.

In order for memory accounting to work correctly, record_batches
and the underlying buffers, must point exactly to the memory they
claim to point and mo more. Otherwise, memory accounting is impossible,
since we wouldn't know which memory the deleters point to.

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200329171010.1188296-1-alex@vectorized.io>