model: support output operator for record batch reader
being able to dump a record batch reader configuration to the log is
important for debugging. in the case of consensus the reader
configuration can be altered in many different ways depending on the
consistency setting and the state of the log. being able to log the
final reader configuration is helpful.

Signed-off-by: Noah Watkins <noah@vectorized.io>