storage: fix reclaimer race with batch cache entries
The synchronous reclaimer creates a challenge for the batch cache which
internally causes allocations to occur which may invoke the reclaimer.

Batch cache operations that hold a reference to an entry in the cache need to
be careful that any heap allocation may cause the held reference to be
_concurrently_ freed/invalidated.

For example, given a live entry in the cache:

   auto e = _index.find(..);
   assert(e);

doing something such as `e->batch.share()` internally causes allocations
which may trigger the reclaimer. if the reclaimer selects `e` for reclaiming,
then `e->batch.share()` will run concurrently with the destructor of the
batch (not technically concurently--but interleaved with share() in ways that
are not safe).

If an operation may perform an allocation use entry::pin/unpin to guard the
reference which will force the reclaimer to skip the entry.

Signed-off-by: Noah Watkins <noah@vectorized.io>