kafka: add parsing support for offset commit version 1
this version adds (and is removed in version 2) a commit timestamp to
the partition. it appears to be involved in expiring timestamps, which
in later versions of kafka seem to be disabled by default. so this patch
adds parsing to support clients that want version 1 of the wire protocol
but further testing is needed to determine what behaviors the clients
want.

Signed-off-by: Noah Watkins <noah@vectorized.io>