kafka: on group join check if members need woken up
before returning from join group request, for existing groups, we check
if all the members have joined, and if so, complete the join which wakes
up waiting members as well as completes the current future in the
returning response.

this is also done in other cases within the state machine, and in kafka
is also explicitly handled in the exact case that this patch addresses.
it seems like a bit of an after thought, but does appear to work.

Signed-off-by: Noah Watkins <noah@vectorized.io>