rpk/tuners/irq: Ignore errors when setting an IRQ's affinity
IRQ SMP affinity is tuned on a best-effort basis. Most IO-APIC
compatible IRQs allow their affinity to be set, but there are
exceptions such as IRQ 0, which is the timer IRQ, and other
IRQs that can be managed by the kernel.
Likewise, if an IRQ isn't marked as IO-APIC-compatible, it
doesn't mean its affinity can't be set. Therefore the errors
are logged but otherwise ignored.

Sources:
 - https://www.kernel.org/doc/Documentation/x86/i386/IO-APIC.txt
 - https://cs.uwaterloo.ca/~brecht/servers/apic/SMP-affinity.txt
 - https://github.com/scylladb/scylla/issues/5268#issuecomment-553067312
 -