storage: reduce filesystem contention on metadata by fallocating
::truncate() causes a synchronization point with .size property
of the inode. This change introduces a fake truncation that relies
on our checksummer to do the right thing in case of crashing.

During truncation we zero out 2 pages and do a read-back of the last
half page but follow it with a write of 0's of up to 2 pages. The idea
there is to catch border-conditions between 2 pages.

In addition to not flushing, we adaptively fallocate 32MB at a time
when the append() would exceed the max size