rpc/transport:semaphore units as an option for sequential dispatch
Raft requires sequential dispatch of RPC requests per each group
follower. We dispatch raft requests from both Heartbeat manager and
Consensus implementations. Those requests are serialized by consensus
_op_sem semaphore. In the RPC layer there are multiple continuations
involved between calling send method and the moment when request get
dispatched to the wire. Seastar scheduler can reorder continuation
execution which in the end may lead to the situation where requests will
not be sent by RPC transport in call order. In order to achieve the
sequential dispatch we have to guarantee that calling `send` and actual
dispatch on wire will be executed under the semaphore.
To achieve this added ability to pass in a shared pointer to
semaphore units vector. The units held in this vector will be kept
alive until the request will be written on wire.

Signed-off-by: Michal Maslanka <michal@vectorized.io>