storage: add compaction_strategy configuration to storage::log
This is an infrastructure patch that is very intrusive as
it wraps the ntp passed into each log as a configuration parameter
so we can extend in the future without so much code churn.

This new class ntp-config is where the low level implementation
of the log [ch56] will receive environmental arguments from the
controller which eventually originate from the kafka create topics
api.