gcb: modifies pipeline so latest vtools is used always
Instead of using the `vtools` installation from the builder images, this
modifies the pipeline so that vtools is installed on the workspace so
that the version of the SHA1 being tested is the one being used. This
allows to not have to worry about having the latest version of
vtools in the builder images.

In addition, the pipeline also installs terraform, as it is required by
the new `vtools fmt tf` command.