c/controller: using duration instead of timepoint in autocreate topics
Create topics request contains a timeout that have to be propagated to
the leader controller. Current solution used hackish timeout calculation
subtracting it from current time. (We can not send absolute time point
to leader as clock may not be synchronized between the nodes).
This caused an overflow on the server side  (when
model::no_timeout was used). Instead of calculating timeout duration
changed autocreate API to accept duration type as timeout parameter.

Signed-off-by: Michal Maslanka <michal@vectorized.io>