r/treewide: using committed offset instead of dirty one
When replicating with leader_ack consistency level it is better to use
log committed_offset instead of dirty one. Thanks to this approach we
change raft internal state (visible to the follower) in flush intervals.
This optimizes recovery which is never dispatched when leader dirty
offset is bigger than follower one, only after committed offset is
updated.

Signed-off-by: Michal Maslanka <michal@vectorized.io>