r/treewide: reverted using committed_offset in raft
We flush every time data is appended on the follower node and lazily on
the leader. We missed that while making raft to use committed_offset
instead of dirty one. This caused unnecessary recoveries as the follower
was often ahead of leader as it returned response containing offset that
was already flushed while leader used stale offset to determine if
recovery is needed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20200417134523.839689-1-michal@vectorized.io>