storage:segment_appender: ensure reading happens in full pages
This is a very subtle change that is important and was causing
our crashes.

First we made unrecoverable crashes exit the app. There is nothing
that we can do if we cannot read half. Similarly, there is nothing
that we can do if we issue a truncation on a WAL and it aborts.
These are unrecoverable errors and previously were simply swalloed
by the service handling tier (raft) which was too far removed
from the filesystem to understand the error.

Typically these errors are `system_error (error system:22, Invalid argument)`
which simply mean -EINVAL. This will not fix future corruption issues
but will ensure that we don't cause corruption ourselves.

Second, we fix the source of the bug which was the alignment
to dma_read. This is subtle. Not only are the requirements on read
destination pointer to be hardware aligned via (posix_memalign())
but we must algo gurantee that we read _at least_ a
`file.disk_read_dma_alignment()`. However, because we are interested
in hydrating the last page for the next _write_, we must issue
the alignment to be that of the disk_**write**_dma_alignment() not the reads.