storage:disk_log_impl: fix concurrency bug with truncation
Fixes a concurrency bug with background dispatch of truncation,
creation, and append found by the opfuzzer.

1. create a log segment foo.log
2. dispatch background deletion of foo.log - and suspend
3. dispatch 2 appends
4. actually write 1 append
5. resume background deletion of foo.log
6. write 1 append ---> bug

The reason for this flow is that you can create
with that sequence of operations the *exact* same log that you had
completely truncated before.

For the time being, we move the background deletion of segments
to the foreground and wait for the futures to resolve.
Given that raft is the only user, this will be fine. In the future
when we have replica-reads, we'll need to rework parts of this
by acquiring locks, etc.