tristate: introduced tristate type
The tristate class express a value that can either be present,
not set or disabled explicitly. As the name suggest the tristate can
have the following three states:
 - Disabled
 - Not set
 - Set

Tristate is helpful for example for configuration parameters saying
that feature can be either disabled, or server defaults should
be applied or it should be configured with requested value.

Signed-off-by: Michal Maslanka <michal@vectorized.io>